refactor v1alpha1 secret cred functions
