+++ apis/datadoghq/v1alpha1/secret.go
@@ -3,17 +3,16 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package datadogagent
+package v1alpha1
 
 import (
 	"os"
 
 	apicommon "github.com/DataDog/datadog-operator/apis/datadoghq/common"
-	datadoghqv1alpha1 "github.com/DataDog/datadog-operator/apis/datadoghq/v1alpha1"
 )
 
-// getKeysFromCredentials returns any key data that need to be stored in a new secret
-func getKeysFromCredentials(credentials *datadoghqv1alpha1.DatadogCredentials) map[string][]byte {
+// GetKeysFromCredentials returns any key data that need to be stored in a new secret
+func GetKeysFromCredentials(credentials *DatadogCredentials) map[string][]byte {
 	data := make(map[string][]byte)
 	// Create secret using DatadogAgent credentials if it exists
 	if credentials.APIKey != "" {
@@ -26,21 +25,21 @@ func getKeysFromCredentials(credentials *datadoghqv1alpha1.DatadogCredentials) m
 	return data
 }
 
-// For each of the API key and app key, check if:
+// CheckAPIKeySufficiency use to check for the API key if:
 // 1. an existing secret is defined, or
 // 2. the corresponding env var is defined (whether in ENC format or not)
 // If either of these is true, the secret is not needed for that particular key.
-func checkAPIKeySufficiency(creds *datadoghqv1alpha1.DatadogCredentials, envVarName string) bool {
+func CheckAPIKeySufficiency(creds *DatadogCredentials, envVarName string) bool {
 	return ((creds.APISecret != nil && creds.APISecret.SecretName != "") ||
 		creds.APIKeyExistingSecret != "" ||
 		os.Getenv(envVarName) != "")
 }
 
-// CheckAppKeySufficiency use to check for the APP key if:
+// CheckAPIKeySufficiency use to check for the APP key if:
 // 1. an existing secret is defined, or
 // 2. the corresponding env var is defined (whether in ENC format or not)
 // If either of these is true, the secret is not needed for that particular key.
-func CheckAppKeySufficiency(creds *datadoghqv1alpha1.DatadogCredentials, envVarName string) bool {
+func CheckAppKeySufficiency(creds *DatadogCredentials, envVarName string) bool {
 	return ((creds.APPSecret != nil && creds.APPSecret.SecretName != "") ||
 		creds.AppKeyExistingSecret != "" ||
 		os.Getenv(envVarName) != "")
